fix(ui): shared clipboard helper with an insecure-context fallback for every copy button
navigator.clipboard does not exist on plain HTTP origins, so the SSH key
copy button did nothing over http://<ip>:3000 and swallowed the error.
One helper in lib/utils now tries the async API and falls back to a
hidden textarea with execCommand, replacing the two inline copies of
that fallback and the silent call sites. The SSH key button and the
device IP toast now say so when both paths fail.

Closes #466
